refactor : 비동기 방식 메서드 분리
